+++ out/commands/LELgenerate.js
@@ -27,8 +27,8 @@ exports.LELgenerate = LELgenerate;
 const vscode = __importStar(require("vscode"));
 const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
-const config = __importStar(require("../config"));
-const scan = __importStar(require("../scan"));
+const scan = __importStar(require("./../scan"));
+const autoload_1 = require("./../autoload");
 /**
  * Scan project subdirectories for .php files and process them for ctreating the localization files
  * @returns
@@ -36,10 +36,8 @@ const scan = __importStar(require("../scan"));
 async function LELgenerate() {
     vscode.window.showInformationMessage('Started extracting localization strings!');
     const localizationStrings = new Set();
-    // Load the config file
-    const configJson = config.loadConfig();
     // Search for PHP files and process them in parallel
-    await scan.searchPhpFiles(config.getRootPath(), localizationStrings, configJson.excludePaths, configJson.excludeGitIgnorePaths);
+    await scan.searchPhpFiles(autoload_1.config.rootPath, localizationStrings, autoload_1.config.excludePaths, autoload_1.config.excludeGitIgnorePaths);
     // Generate JSON content
     const jsonContent = {};
     localizationStrings.forEach(str => {
@@ -48,7 +46,7 @@ async function LELgenerate() {
     // Get language codes from user
     const languageCodes = await vscode.window.showInputBox({
         prompt: 'Enter the 2-letter language codes separated by commas (e.g., en,fr,es)',
-        value: configJson.defaultLanguages,
+        value: autoload_1.config.defaultLanguages,
         placeHolder: 'en,fr,es'
     });
     if (!languageCodes) {
@@ -57,16 +55,15 @@ async function LELgenerate() {
     }
     const languages = languageCodes.split(',').map(code => code.trim());
     // Write to lang/{language}.json files
-    const langFolderPath = await config.getLocalizationPath(configJson);
-    if (!langFolderPath) {
+    if (autoload_1.config.langFolderPath === '') {
         vscode.window.showErrorMessage('No localization files generated: localization path not found or could not be determined.');
         return;
     }
-    if (!fs.existsSync(langFolderPath)) {
-        fs.mkdirSync(langFolderPath);
+    if (!fs.existsSync(autoload_1.config.langFolderPath)) {
+        fs.mkdirSync(autoload_1.config.langFolderPath);
     }
     for (const lang of languages) {
-        const outputFilePath = path.join(langFolderPath, `${lang}.json`);
+        const outputFilePath = path.join(autoload_1.config.langFolderPath, `${lang}.json`);
         // Check if file exists and merge with existing content
         let existingContent = {};
         if (fs.existsSync(outputFilePath)) {