+++ out/config.js
@@ -71,7 +71,8 @@ function loadConfig() {
         excludePaths: [],
         excludeGitIgnorePaths: true,
         autoDetectLocalizationPath: false,
-        localizationPath: ""
+        localizationPath: "",
+        disableCache: false
     };
 }
 /**
@@ -117,6 +118,7 @@ function normalizeConfig(config, rootPath) {
     config.excludeGitIgnorePaths = (config.excludeGitIgnorePaths ?? true); // Default true if not present
     config.autoDetectLocalizationPath = (config.autoDetectLocalizationPath ?? false); // Default false if not present
     config.localizationPath = (config.localizationPath ?? "");
+    config.disableCache = (config.disableCache ?? false);
 }
 /**
  * Determines the language directory path based on Laravel's version and common structure.